Test using mainnet blocks
