+++ parachain/pallets/verifier-lightclient/Cargo.toml
@@ -12,6 +12,10 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 hex = { package = "rustc-hex", version = "2.1.0", default-features = false }
 
+[dev-dependencies]
+serde_json = "1.0"
+artemis-testutils = { path = "../../primitives/testutils" }
+
 [dependencies.codec]
 default-features = false
 features = ["derive"]
