+++ parachain/pallets/verifier-lightclient/src/testdata/11090290.json
@@ -1,0 +1,20 @@
+{
+    "difficulty": "0xbc140caa61087",
+    "extra_data": [101, 116, 104, 101, 114, 109, 105, 110, 101, 45, 97, 115, 105, 97, 49, 45, 51],
+    "gas_limit": "0xbe8c19",
+    "gas_used": "0x0",
+    "hash": "0x0f9bdc91c2e0140acb873330742bda8c8181fa3add91fe7ae046251679cedef7",
+    "logs_bloom": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+    "author": "0xea674fdde714fd979de3edf0f56aa9716b898ec8",
+    "number": 11090290,
+    "parent_hash": "0xbede0bddd6f32c895fc505ffe0c39d9bde58e9a5272f31a3dee448b796edcbe3",
+    "receipts_root": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+    "ommers_hash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
+    "seal": [
+        [160, 190, 58, 223, 176, 8, 123, 230, 43, 40, 183, 22, 226, 205, 243, 199, 147, 41, 223, 92, 170, 4, 201, 238, 224, 53, 211, 91, 93, 82, 16, 40, 21],
+        [136, 105, 53, 187, 231, 182, 60, 79, 142]
+    ],
+    "state_root": "0x7dcb8aca872b712bad81df34a89d4efedc293566ffc3eeeb5cbcafcc703e42c9",
+    "timestamp": 1603160977,
+    "transactions_root": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421"
+}
