+++ parachain/pallets/verifier-lightclient/src/testdata/11090292.json
@@ -1,0 +1,20 @@
+{
+    "difficulty": "0xbc43149e5bdb5",
+    "extra_data": [112, 112, 121, 101, 32, 112, 97, 110, 100, 97, 95, 112, 111, 111, 108],
+    "gas_limit": "0xbe8c0e",
+    "gas_used": "0xbe7ca2",
+    "hash": "0xb25205d51d3f71c2688c7490458c0df120e351d604ac0e4ddcfc28751a14faa8",
+    "logs_bloom": [84, 35, 76, 232, 65, 96, 234, 170, 212, 66, 62, 118, 174, 27, 107, 4, 200, 128, 160, 242, 141, 45, 154, 181, 28, 73, 2, 86, 40, 146, 7, 100, 244, 222, 208, 72, 22, 138, 86, 34, 216, 184, 170, 67, 10, 173, 71, 248, 30, 96, 92, 58, 75, 98, 8, 207, 137, 234, 90, 192, 225, 36, 129, 104, 121, 137, 42, 5, 147, 210, 20, 136, 72, 78, 90, 108, 13, 145, 16, 170, 78, 85, 244, 116, 18, 117, 81, 138, 90, 117, 45, 76, 131, 246, 155, 8, 126, 132, 3, 144, 14, 11, 161, 165, 123, 203, 23, 2, 130, 245, 216, 22, 96, 8, 130, 40, 4, 47, 223, 167, 102, 20, 36, 183, 18, 54, 216, 64, 25, 109, 155, 101, 133, 84, 181, 80, 122, 106, 136, 195, 11, 4, 41, 91, 64, 23, 109, 181, 113, 84, 147, 57, 150, 73, 35, 69, 141, 81, 46, 149, 94, 41, 134, 138, 27, 100, 160, 66, 102, 114, 1, 158, 1, 82, 64, 113, 8, 18, 68, 150, 40, 2, 8, 56, 231, 78, 62, 66, 228, 136, 200, 31, 168, 12, 133, 15, 57, 79, 166, 3, 1, 4, 194, 210, 112, 107, 76, 201, 33, 224, 50, 18, 28, 146, 10, 94, 208, 43, 113, 58, 52, 98, 241, 140, 65, 81, 100, 27, 145, 8, 193, 69, 10, 192, 6, 64, 21, 27, 46, 99, 164, 74, 252, 201, 197, 48, 197, 196, 236, 201, 102, 102, 106, 128, 25, 80],
+    "author": "0x2a5994b501e6a560e727b6c2de5d856396aadd38",
+    "number": 11090292,
+    "parent_hash": "0xb5500f8ba8e33719909eb686c2ceefc175bcacd87e24544f04dda72451797da9",
+    "receipts_root": "0xddee101090a7a85b2e515060bee5cbbde552057d65e3e3c04b327eddd1ea81ca",
+    "ommers_hash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
+    "seal": [
+        [160, 79, 100, 120, 109, 190, 51, 29, 60, 226, 218, 29, 12, 10, 225, 255, 128, 12, 59, 152, 181, 25, 104, 110, 223, 61, 104, 216, 46, 26, 88, 68, 139],
+        [136, 225, 237, 252, 190, 210, 24, 159, 128]
+    ],
+    "state_root": "0xbe8a8544bba60fb8c995a51510530e1dd0e9d1098dfcf201d6e795528d41022d",
+    "timestamp": 1603160981,
+    "transactions_root": "0xfab82306ea3b6e853c1f8c5fc967b14d1e38473b0c755226920f657943d1acde"
+}
