+++ parachain/pallets/verifier-lightclient/src/tests.rs
@@ -1,12 +1,15 @@
+#![cfg(test)]
+
 use crate::mock::{
-	child_of_genesis_ethereum_header, new_tester, AccountId, Verifier, MockRuntime, Origin,
+	child_of_genesis_ethereum_header, ethereum_header_from_file,
+	ethereum_header_proof_from_file, new_tester, new_tester_with_config,
+	AccountId, Verifier, MockRuntime, Origin,
 };
 use crate::sp_api_hidden_includes_decl_storage::hidden_include::{StorageMap, StorageValue};
 use frame_support::{assert_err, assert_ok};
 use sp_keyring::AccountKeyring as Keyring;
 use sp_runtime::DispatchError;
-
-use crate::{Error, EthereumHeader, BestBlock, Headers, HeadersByNumber};
+use crate::{BestBlock, Error, EthereumHeader, GenesisConfig, Headers, HeadersByNumber};
 
 #[test]
 fn it_tracks_highest_difficulty_ethereum_chain() {
@@ -18,8 +21,16 @@ fn it_tracks_highest_difficulty_ethereum_chain() {
 		child2.difficulty = 0x20000.into();
 
 		let ferdie: AccountId = Keyring::Ferdie.into();
-		assert_ok!(Verifier::import_header(Origin::signed(ferdie.clone()), child1));
-		assert_ok!(Verifier::import_header(Origin::signed(ferdie.clone()), child2));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			child1,
+			Default::default(),
+		));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			child2,
+			Default::default(),
+		));
 	
 		let (header_id, highest_difficulty) = BestBlock::get();
 		assert_eq!(header_id.hash, child1_hash);
@@ -39,8 +50,16 @@ fn it_tracks_multiple_unconfirmed_ethereum_forks() {
 		let child2_hash = child2.compute_hash();
 
 		let ferdie: AccountId = Keyring::Ferdie.into();
-		assert_ok!(Verifier::import_header(Origin::signed(ferdie.clone()), child1));
-		assert_ok!(Verifier::import_header(Origin::signed(ferdie.clone()), child2));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			child1,
+			Default::default(),
+		));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			child2,
+			Default::default(),
+		));
 
 		assert!(Headers::<MockRuntime>::contains_key(child1_hash));
 		assert!(Headers::<MockRuntime>::contains_key(child2_hash));
@@ -56,9 +75,17 @@ fn it_imports_ethereum_header_only_once() {
 		let child_for_reimport = child.clone();
 
 		let ferdie: AccountId = Keyring::Ferdie.into();
-		assert_ok!(Verifier::import_header(Origin::signed(ferdie.clone()), child));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			child,
+			Default::default(),
+		));
 		assert_err!(
-			Verifier::import_header(Origin::signed(ferdie.clone()), child_for_reimport),
+			Verifier::import_header(
+				Origin::signed(ferdie.clone()),
+				child_for_reimport,
+				Default::default(),
+			),
 			Error::<MockRuntime>::DuplicateHeader,
 		);
 	});
@@ -69,7 +96,7 @@ fn it_rejects_unsigned_ethereum_header() {
 	new_tester().execute_with(|| {
 		let child = child_of_genesis_ethereum_header();
 		assert_err!(
-			Verifier::import_header(Origin::none(), child),
+			Verifier::import_header(Origin::none(), child, Default::default()),
 			DispatchError::BadOrigin,
 		);
 	});
@@ -84,8 +111,38 @@ fn it_rejects_ethereum_header_before_parent() {
 
 		let ferdie: AccountId = Keyring::Ferdie.into();
 		assert_err!(
-			Verifier::import_header(Origin::signed(ferdie), child_of_child),
+			Verifier::import_header(
+				Origin::signed(ferdie),
+				child_of_child,
+				Default::default(),
+			),
 			Error::<MockRuntime>::MissingParentHeader,
 		);
+	});
+}
+
+#[test]
+fn it_validates_proof_of_work() {
+	new_tester_with_config(GenesisConfig {
+		initial_header: ethereum_header_from_file(11090290),
+		initial_difficulty: 0.into(),
+		verify_pow: true,
+	}).execute_with(|| {
+		let header1 = ethereum_header_from_file(11090291);
+		let header1_proof = ethereum_header_proof_from_file(11090291);
+		let header2 = ethereum_header_from_file(11090292);
+		let header2_proof = ethereum_header_proof_from_file(11090292);
+
+		let ferdie: AccountId = Keyring::Ferdie.into();
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie.clone()),
+			header1,
+			header1_proof,
+		));
+		assert_ok!(Verifier::import_header(
+			Origin::signed(ferdie),
+			header2,
+			header2_proof,
+		));
 	});
 }
