+++ parachain/primitives/ethereum/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false }
 ethbloom = { version = "0.8", default-features = false }
-ethereum-types = { version = "0.9.2", default-features = false }
+ethereum-types = { version = "0.9.2", default-features = false, features = ["codec"] }
 hex = { package = "rustc-hex", version = "2.1.0", default-features = false }
 hex-literal = { version = "0.3.1", default-features = false }
 libsecp256k1 = { version = "0.3", default-features = false }
@@ -73,6 +73,7 @@ std = [
 ]
 
 [dev-dependencies]
+artemis-testutils = { path = "../../primitives/testutils" }
 rand = "0.7.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
