+++ parachain/primitives/testutils/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "artemis-testutils"
+version = "0.1.0"
+authors = ["Snowfork <contact@snowfork.com>"]
+edition = "2018"
+
+[dependencies]
+ethereum-types = { version = "0.9.2", default-features = false }
+hex = { package = "rustc-hex", version = "2.1.0", default-features = false }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
